+++ database/migrations/2023_04_14_104036_create_appointments_daily_stats_table.php
@@ -15,13 +15,13 @@ class CreateAppointmentsDailyStatsTable extends Migration
     {
         Schema::create('appointments_daily_stats', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('centre_id');
+            $table->unsignedBigInteger('centre_id');
             $table->foreign('centre_id')->references('id')->on('locations');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->unsignedInteger('appointment_id');
+            $table->unsignedBigInteger('appointment_id');
             $table->foreign('appointment_id')->references('id')->on('appointments');
-            $table->unsignedInteger('appointment_status_id');
+            $table->unsignedBigInteger('appointment_status_id');
             $table->foreign('appointment_status_id')->references('id')->on('appointment_statuses');
             $table->date('cron_current_date');
             $table->timestamps();
